Fixup more signed/unsigned comparisons.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@3790 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmpv3.c
@@ -566,7 +566,7 @@ init_snmpv3_post_config(int majorid, int minorid, void *serverarg,
   }
 
   /* if our engineID has changed at all, the boots record must be set to 1 */
-  if (engineIDLen != oldEngineIDLength ||
+  if (engineIDLen != (int)oldEngineIDLength ||
       oldEngineID == NULL || c_engineID == NULL ||
       memcmp(oldEngineID, c_engineID, engineIDLen) != 0) {
     engineBoots = 1;
